move types into global type file, consolidate some types, remove unused types
