+++ src/content/hooks/useBugFetcher.ts
@@ -1,57 +1,10 @@
 import { useState, useEffect } from "react";
 import { runCachedQueries } from "../lib/utils";
 
-export interface Bug {
-  [key: string]: any;
-}
-
-export interface Rule {
-  key: string;
-  operator: string;
-  value: string;
-}
-
-export interface BugsByMeta {
-  [key: string]: any;
-}
-
-export interface RuleSet {
-  operator?: "OR" | "AND";
-  rules: Array<Rule | RuleSet>;
-}
-
-export interface BugQuery {
-  include_fields: string[];
-  resolution?: string;
-  priority?: string;
-  order?: string;
-  component?: string[];
-  rules: Array<Rule | RuleSet>;
-}
-
-export interface UseBugFetcherOptions {
-  query: BugQuery;
-  updateOn?: any[];
-  bugsByMeta?: any[];
-  transformBugs?: (bugs: Bug[]) => Bug[];
-}
-
 export interface UseBugFetcherReturn {
   status: "" | "loading" | "loaded";
   bugs: Bug[];
   bugsByMeta?: BugsByMeta[];
-  awaitingNetwork: boolean;
-}
-
-type BugQueryResponse = { bugs: Bug[] };
-
-export interface BugQueryReturn {
-  rsp: BugQueryResponse;
-  awaitingNetwork: boolean;
-}
-
-export interface BugQueriesReturn {
-  rsp: Array<BugQueryResponse>;
   awaitingNetwork: boolean;
 }
